requirements.txt
uvicorn~=0.23.2
fastapi~=0.103.2
SQLAlchemy~=2.0.21
pydantic~=2.4.2
python-dotenv~=1.0.0
passlib~=1.7.4
jose~=1.0.0